refactoring and fix norms

+++ includes/point.h
@@ -6,14 +6,14 @@
 /*   By: ybutkov <ybutkov@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/21 13:17:14 by ybutkov           #+#    #+#             */
-/*   Updated: 2025/09/22 19:51:06 by ybutkov          ###   ########.fr       */
+/*   Updated: 2025/09/28 20:03:43 by ybutkov          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef POINT_H
 # define POINT_H
 
-#include <stdint.h>
+# include <stdint.h>
 
 typedef int16_t		t_z_type;
 typedef uint32_t	t_color_type;
@@ -27,15 +27,15 @@ typedef struct s_point_2d
 
 typedef struct s_point_3d
 {
-	int				x;
-	int				y;
-	int				z;
+	double			x;
+	double			y;
+	double			z;
 }					t_point_3d;
 
 typedef struct s_point
 {
 	t_z_type		z;
 	t_color_type	color;
-} __attribute__((packed)) t_point;
+} __attribute__((packed))	t_point;
 
 #endif
